untangle surface water functions
